+++ packages/terra-date-picker/CHANGELOG.md
@@ -7,6 +7,10 @@
   * Removed selected date information on focus of picker button.
   * Added ariaDescribedBy prop in `terra-date-time-picker` example for invalid error message.
 
+* Changed
+  * Fixed the invalid case for SR announcement.
+  * Changed the SR announcement for changing date from calender.
+
 ## 4.98.1 - (February 16, 2024)
 
 * Fixed